fix size_t warning

+++ rts/System/SpringApp.cpp
@@ -827,7 +827,7 @@ void SpringApp::Reload(const std::string script)
 	// clean changed configs
 	configHandler->Update();
 
-	LOG("[SpringApp::%s][12] #script=%lu", __func__, script.size());
+	LOG("[SpringApp::%s][12] #script=%lu", __func__, (unsigned long) script.size());
 
 	if (script.empty()) {
 		// if no script, drop back to menu
